feature: Add an special to string method to types

+++ tests/error/comptime_expr_errors/unsupported_in_list.py
@@ -5,7 +5,7 @@ from guppylang.std.builtins import comptime
 from guppylang.std.qsystem.utils import get_current_shot
 from guppylang.std.quantum import qubit, discard
 
-randints = list(np.random.choice(3, 50, p = [0.1, 0.5, 0.4]))
+randints = list(np.random.choice(3, 50, p = [0.1, 0.5, 0.4]).astype(np.int64))
 
 @guppy
 def main() -> None:
